Move test to `test_transform`
